PageForm: extract SlugField
